Refactor country page to use CountryPageContent component
Moved country data fetching and rendering logic from page.tsx to a new CountryPageContent component for better separation of concerns and improved code organization.
